Add files via upload
This is an AI model inspired by ChatGPT, designed for natural language understanding and conversational tasks. Built using transformer-based architecture, it generates human-like responses, engages in meaningful dialogues, and can be fine-tuned for specific domains. The model supports a variety of use cases, from answering queries to text generation. Easy to use and customizable, this project is ideal for developers interested in creating AI-powered chatbots or conversational agents. 

Javascript code is not available due to privacy reasons . If You want then message me or connect me on liknedin.